requirements.txt
Django==5.2.7
psycopg2-binary==2.9.9
python-decouple==3.8
django-environ==0.11.2
python-dotenv==1.0.0
whitenoise==6.6.0
gunicorn==21.2.0
reportlab
openpyxl
django-filter
djangorestframework
djangorestframework-simplejwt
django-cors-headers

# Supabase (Opcional - para uso futuro)
# supabase==2.3.0